Change Id to int
어차피 유저나 리포나 1억개가 안됨.
int일 필요가 없음.

+++ cmd/ga_crawler/event.go
@@ -5,12 +5,12 @@ import (
 )
 
 type User struct {
-	Id    uint64 `json:"id"`
+	Id    int    `json:"id"`
 	Login string `json:"login"`
 }
 
 type Repo struct {
-	Id   uint64 `json:"id"`
+	Id   int    `json:"id"`
 	Name string `json:"name"`
 }
